some issues fixed: device doesnt activated, get pcb serial number

+++ certoclavLafomed/src/main/java/com/certoclav/app/database/Profile.java
@@ -277,7 +277,7 @@ public class Profile {
                    boolean isMaintainEnabled,
                    boolean isContByFlexProbe1,
                    boolean isContByFlexProbe2,
-                   int finalTemp,
+                   float finalTemp,
                    float f0Value,
                    float zValue) {
         this.cloudId = cloudId;
